% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{check_dir}
\alias{check_dir}
\title{Set up the directory structure for the checks}
\usage{
check_dir(pkgdir, what = c("check", "db", "old", "new", "pkg"),
  package = NULL)
}
\arguments{
\item{pkgdir}{Path to the package we are revdep-checking.}

\item{what}{Directory to query:
\itemize{
\item \code{"check"}: the root of the check directory,
\item \code{"db"}: the database file,
\item \code{"old"}: the library of the old version of the package.
\item \code{"new"}: the library of the new version of the package.
\item \code{"pkg"}: the library of the reverse dependency, the \code{package}
argument must be supplied as well.
}}

\item{package}{The name of the package, if \code{what} is \code{"pkg"}.}
}
\value{
Character scalar, the requested path.
}
\description{
Currently the following files and directories are used.
They are all in the main revdep directory, which is \code{revdep} in the
package tree.
\itemize{
\item \code{library}: a collection of package libraries
\item \code{library/data.sqlite}: the SQLite database that contains the check
data.
\item \code{library/<checked-pkg>/old}: library that contains the \emph{old} version
of the revdep-checked package, together with its dependencies.
\item \code{library/<checked-pkg>/new}: library that contains the \emph{new} version
of the revdep-checked package, together with its dependencies.
\item \code{library/<pkg>} are the libraries for the reverse dependencies.
}
}
\keyword{internal}
