% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_broken}
\alias{cloud_broken}
\title{Retrieve the names broken packages}
\usage{
cloud_broken(
  job_id = cloud_job(),
  pkg = ".",
  install_failures = FALSE,
  timeout_failures = FALSE
)
}
\arguments{
\item{job_id}{The batch job_id, as returned by \code{\link[=cloud_check]{cloud_check()}}.}

\item{pkg}{Path to package.}

\item{install_failures}{Whether to include packages that failed to install.}

\item{timeout_failures}{Whether to include packages that timed out.}
}
\value{
A character vector with the names of broken packages, to be passed to \code{cloud_check()}.
}
\description{
Retrieve the names broken packages
}
\seealso{
Other cloud: 
\code{\link{cloud_cancel}()},
\code{\link{cloud_check}()},
\code{\link{cloud_details}()},
\code{\link{cloud_fetch_results}()},
\code{\link{cloud_plot}()},
\code{\link{cloud_report}()},
\code{\link{cloud_results}()},
\code{\link{cloud_status}()},
\code{\link{cloud_summary}()}
}
\concept{cloud}
