% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_browse}
\alias{cloud_browse}
\title{Browse to the AWS url for the job}
\usage{
cloud_browse(job_id = cloud_job())
}
\arguments{
\item{job_id}{The batch job_id, as returned by \code{\link[=cloud_check]{cloud_check()}}.}
}
\description{
This is useful for closer inspection of individual jobs while they are
running or after the fact.
}
