% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_cancel}
\alias{cloud_cancel}
\title{Cancel a running cloud run}
\usage{
cloud_cancel(job_name = cloud_job())
}
\arguments{
\item{job_name}{The job name, as returned by \code{\link[=cloud_check]{cloud_check()}}.}
}
\description{
Cancel a running cloud run
}
\seealso{
Other cloud: 
\code{\link{cloud_broken}()},
\code{\link{cloud_check}()},
\code{\link{cloud_details}()},
\code{\link{cloud_fetch_results}()},
\code{\link{cloud_plot}()},
\code{\link{cloud_report}()},
\code{\link{cloud_results}()},
\code{\link{cloud_status}()},
\code{\link{cloud_summary}()}
}
\concept{cloud}
