% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_check}
\alias{cloud_check}
\title{Submit a reverse dependency checking job to the cloud}
\usage{
cloud_check(
  pkg = ".",
  tarball = NULL,
  revdep_packages = NULL,
  r_version = "4.1.1",
  check_args = "--no-manual"
)
}
\arguments{
\item{pkg}{Path to package.}

\item{tarball}{A pre-built package tarball, if \code{NULL} a tarball will be
automatically built for the package at \code{pkg} by \code{\link[pkgbuild:build]{pkgbuild::build()}}.}

\item{revdep_packages}{A character vector of packages to check, if \code{NULL}
equal to \code{\link[=cran_revdeps]{cran_revdeps()}}}

\item{r_version}{The R version to use.}

\item{check_args}{Additional argument to pass to \verb{R CMD check}}
}
\value{
The AWS Batch job-id
}
\description{
Submit a reverse dependency checking job to the cloud
}
\seealso{
Other cloud: 
\code{\link{cloud_broken}()},
\code{\link{cloud_cancel}()},
\code{\link{cloud_details}()},
\code{\link{cloud_fetch_results}()},
\code{\link{cloud_plot}()},
\code{\link{cloud_report}()},
\code{\link{cloud_results}()},
\code{\link{cloud_status}()},
\code{\link{cloud_summary}()}
}
\concept{cloud}
