% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_details}
\alias{cloud_details}
\title{Display detailed revdep results from a cloud run}
\usage{
cloud_details(job_name = cloud_job(pkg = pkg), revdep, pkg = ".")
}
\arguments{
\item{job_name}{The job name, as returned by \code{\link[=cloud_check]{cloud_check()}}.}

\item{revdep}{Name of the revdep package}

\item{pkg}{Path to package.}
}
\description{
Display detailed revdep results from a cloud run
}
\seealso{
Other cloud: 
\code{\link{cloud_broken}()},
\code{\link{cloud_cancel}()},
\code{\link{cloud_check}()},
\code{\link{cloud_fetch_results}()},
\code{\link{cloud_plot}()},
\code{\link{cloud_report}()},
\code{\link{cloud_results}()},
\code{\link{cloud_status}()},
\code{\link{cloud_summary}()}
}
\concept{cloud}
