% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_fetch_results}
\alias{cloud_fetch_results}
\title{Fetch results from the cloud}
\usage{
cloud_fetch_results(job_name = cloud_job(pkg = pkg), pkg = ".")
}
\arguments{
\item{job_name}{The job name, as returned by \code{\link[=cloud_check]{cloud_check()}}.}

\item{pkg}{Path to package.}
}
\description{
Intended mainly for internal and expert use. This function when needed by
\code{\link[=cloud_report]{cloud_report()}} and \verb{[cloud_summary()]}, so it is unlikely you will need to
call it explicitly.
}
\seealso{
Other cloud: 
\code{\link{cloud_broken}()},
\code{\link{cloud_cancel}()},
\code{\link{cloud_check}()},
\code{\link{cloud_details}()},
\code{\link{cloud_plot}()},
\code{\link{cloud_report}()},
\code{\link{cloud_results}()},
\code{\link{cloud_status}()},
\code{\link{cloud_summary}()}
}
\concept{cloud}
\keyword{internal}
