% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_plot}
\alias{cloud_plot}
\title{Plot the running time per package of a cloud job}
\usage{
cloud_plot(job_name = cloud_job())
}
\arguments{
\item{job_name}{The job name, as returned by \code{\link[=cloud_check]{cloud_check()}}.}
}
\description{
Plot the running time per package of a cloud job
}
\seealso{
Other cloud:
\code{\link[=cloud_broken]{cloud_broken()}},
\code{\link[=cloud_browse]{cloud_browse()}},
\code{\link[=cloud_cancel]{cloud_cancel()}},
\code{\link[=cloud_check]{cloud_check()}},
\code{\link[=cloud_details]{cloud_details()}},
\code{\link[=cloud_fetch_results]{cloud_fetch_results()}},
\code{\link[=cloud_report]{cloud_report()}},
\code{\link[=cloud_results]{cloud_results()}},
\code{\link[=cloud_status]{cloud_status()}},
\code{\link[=cloud_summary]{cloud_summary()}}
}
\concept{cloud}
