% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_report}
\alias{cloud_report}
\alias{cloud_report_summary}
\alias{cloud_report_problems}
\alias{cloud_report_failures}
\alias{cloud_report_cran}
\title{Markdown report of reverse dependency check results from the cloud}
\usage{
cloud_report(
  job_name = cloud_job(pkg = pkg),
  pkg = ".",
  file = "",
  all = FALSE,
  results = NULL,
  failures = TRUE
)

cloud_report_summary(
  job_name = cloud_job(pkg = pkg),
  file = "",
  all = FALSE,
  pkg = ".",
  results = NULL
)

cloud_report_problems(
  job_name = cloud_job(pkg = pkg),
  pkg = ".",
  file = "",
  all = FALSE,
  results = NULL
)

cloud_report_failures(
  job_name = cloud_job(pkg = pkg),
  pkg = ".",
  file = "",
  results = NULL
)

cloud_report_cran(job_name = cloud_job(pkg = pkg), pkg = ".", results = NULL)
}
\arguments{
\item{job_name}{The job name, as returned by \code{\link[=cloud_check]{cloud_check()}}.}

\item{pkg}{Path to package.}

\item{file}{File to write output to. Default will write to console.}

\item{all}{Whether to report all problems, including the ones that
were already present in the old version of the package. This potentially
generated a lot of output, most of which was irrelevant, so they are
omitted by default, and only problems seen with the new version of
the package are reported.}

\item{results}{Results from \code{\link[=cloud_results]{cloud_results()}}. Expert use only.}

\item{failures}{Save failures to disk?}
}
\description{
You can use these functions to get intermediate reports of a running cloud check.
}
\seealso{
Other cloud: 
\code{\link{cloud_broken}()},
\code{\link{cloud_cancel}()},
\code{\link{cloud_check}()},
\code{\link{cloud_details}()},
\code{\link{cloud_fetch_results}()},
\code{\link{cloud_plot}()},
\code{\link{cloud_results}()},
\code{\link{cloud_status}()},
\code{\link{cloud_summary}()}
}
\concept{cloud}
