% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_results}
\alias{cloud_results}
\title{Retrieve cloud results}
\usage{
cloud_results(job_name = cloud_job(pkg = pkg), pkg = ".")
}
\arguments{
\item{job_name}{The job name, as returned by \code{\link[=cloud_check]{cloud_check()}}.}

\item{pkg}{Path to package.}
}
\description{
Intended for expert use only, this can be used as input to the \code{\link[=cloud_report]{cloud_report()}} and other functions.
}
\seealso{
Other cloud: 
\code{\link{cloud_broken}()},
\code{\link{cloud_browse}()},
\code{\link{cloud_cancel}()},
\code{\link{cloud_check}()},
\code{\link{cloud_details}()},
\code{\link{cloud_fetch_results}()},
\code{\link{cloud_plot}()},
\code{\link{cloud_report}()},
\code{\link{cloud_status}()},
\code{\link{cloud_summary}()}
}
\concept{cloud}
\keyword{internal}
