% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_summary}
\alias{cloud_summary}
\title{Display revdep results}
\usage{
cloud_summary(job_name = cloud_job(pkg = pkg), pkg = ".")
}
\arguments{
\item{job_name}{The job name, as returned by \code{\link[=cloud_check]{cloud_check()}}.}

\item{pkg}{Path to package.}
}
\description{
Displays nicely formatted results of processed packages run in the cloud.
}
\seealso{
Other cloud:
\code{\link[=cloud_broken]{cloud_broken()}},
\code{\link[=cloud_browse]{cloud_browse()}},
\code{\link[=cloud_cancel]{cloud_cancel()}},
\code{\link[=cloud_check]{cloud_check()}},
\code{\link[=cloud_details]{cloud_details()}},
\code{\link[=cloud_fetch_results]{cloud_fetch_results()}},
\code{\link[=cloud_plot]{cloud_plot()}},
\code{\link[=cloud_report]{cloud_report()}},
\code{\link[=cloud_results]{cloud_results()}},
\code{\link[=cloud_status]{cloud_status()}}
}
\concept{cloud}
