% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{cran_compare}
\alias{cran_compare}
\alias{async_check_package}
\title{Compare local checks to CRAN checks}
\usage{
async_check_package(dir, pkg)
}
\arguments{
\item{dir}{The directory to perform local checks in.}

\item{pkg_name}{A package name.}

\item{flavour_pattern}{A regexp to match against
\code{\link[rcmdcheck:cran_check_flavours]{rcmdcheck::cran_check_flavours()}}. If multiple matches are
found, the first element is used. This determines which CRAN
check to compare against.}
}
\description{
Compare local checks against CRAN checks. This can be useful for
running checks on a local patched version of R with the CRAN
results for that package. For this reason, the default value of
\code{flavour_pattern} is \code{"devel"}
}
