% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirs.R
\name{dir_find}
\alias{dir_find}
\alias{dir_setup}
\alias{dir_setup_package}
\title{Set up/retrieve the directory structure for the checks}
\usage{
dir_find(
  pkgdir,
  what = c("root", "db", "old", "new", "pkg", "check", "checks", "lib", "pkgold",
    "pkgnew", "cloud", "tools", "cache"),
  package = NULL
)

dir_setup(pkgdir)

dir_setup_package(pkgdir, package)
}
\arguments{
\item{pkgdir}{Path to the package we are revdep-checking, or a root
created by \code{\link[=rdev_init]{rdev_init()}}.}

\item{what}{Directory to query:
\itemize{
\item \code{"root"}: the root of the check directory,
\item \code{"db"}: the database file,
\item \code{"old"}: the library of the old version of the package.
\item \code{"new"}: the library of the new version of the package.
\item \code{"pkg"}: the library of the reverse dependency, the \code{package}
argument must be supplied as well.
\item \code{"check"}: the check directory of the reverse dependency, the
\code{package} argument must be supplied as well.
\item \code{"pkgold"}: package libraries to use when checking \code{package} with
the old version.
\item \code{"pkgnew"}: package libraries to use when checking \code{package} with
the new version.
\item \code{"tools"}: the tools library, see above.
\item \code{"cache"}: the default crancache cache directory, see above.
}

An \code{\link[=rdev_init]{rdev_init()}} root has no package under test, so \code{"old"} and
\code{"new"} are \code{NULL} there, and \code{"pkgold"} and \code{"pkgnew"} are just the
library of the reverse dependency.}

\item{package}{The name of the package, if \code{what} is \code{"pkg"}, \code{"check"},
\code{"pkgold"} or \code{"pkgnew"}.}
}
\value{
Character scalar, the requested path.
}
\description{
Currently the following files and directories are used.
They are all in the main revdep directory, which is \code{revdep} in the
package tree, or the root directory itself for a check set up with
\code{\link[=rdev_init]{rdev_init()}}.
\itemize{
\item \code{library}: a collection of package libraries
\item \code{data.sqlite}: the SQLite database that contains the check data.
\item \verb{library/<checked-pkg>/old}: library that contains the \emph{old} version
of the revdep-checked package, together with its dependencies.
\item \verb{library/<checked-pkg>/new}: library that contains the \emph{new} version
of the revdep-checked package, together with its dependencies.
\item \verb{library/<pkg>} are the libraries for the reverse dependencies.
\item \code{tools}: library with the packages the worker processes need when
they run a different R build than the current session, see
\code{\link[=rdev_check]{rdev_check()}}.
\item \code{crancache}: the default crancache package cache of \code{\link[=rdev_check]{rdev_check()}}.
}
}
\keyword{internal}
