% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ revdepcheck }
\alias{ revdepcheck }
\alias{ revdepcheck }
\title{ Automated Reverse Dependency Checking }
\description{
Automated, isolated reserve dependency checking, with automatic comparison of
  the results to the current CRAN checks.
}
