% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/todo.R
\name{revdep_add}
\alias{revdep_add}
\alias{revdep_add_broken}
\alias{revdep_add_new}
\alias{revdep_todo}
\alias{revdep_rm}
\title{Manage the package checking to-do list.}
\usage{
revdep_add(pkg = ".", packages)

revdep_add_broken(pkg = ".", install_failures = FALSE,
  timeout_failures = FALSE)

revdep_add_new(pkg = ".")

revdep_todo(pkg = ".")

revdep_rm(pkg = ".", packages)
}
\arguments{
\item{pkg}{Path to package.}

\item{packages}{Character vector of package names to add}

\item{install_failures}{Whether to re-add packages that failed to
install.}

\item{timeout_failures}{Whether to re-add packages that timed out.}
}
\description{
\code{revdep_todo()} tells you which packages still need to be checked.
\code{revdep_add()} adds a single package to the to-do list.
\code{revdep_rm()} removes packages from the todo list.
\code{revdep_add_broken()} re-adds all broken packages from the last check
(this is useful if you think you've fixed the underlying problem in
your package).
}
