% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{revdep_details}
\alias{revdep_details}
\title{Display details of a single revdepcheck}
\usage{
revdep_details(pkg = ".", revdep)
}
\arguments{
\item{pkg}{Path to package}

\item{revdep}{Name of revdep package.}
}
\description{
Use this to see nicely formatted results of processed packages while
\code{\link[=revdep_check]{revdep_check()}} is running in another process.
}
