% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{revdep_email}
\alias{revdep_email}
\alias{revdep_email_draft}
\title{Notify revdep maintainers about problems}
\usage{
revdep_email(pkg = ".")

revdep_email_draft(pkg, data = email_data(pkg))
}
\arguments{
\item{pkg}{Path to package}

\item{data}{Optionally, supply a named list to provide your own parameters
to fill in the template}
}
\description{
This function uses gmail to automatically notify all maintainers of revdeps
that have failures with the new version of the package. The form of the
email is fixed, but it uses template parameters so that you can control
the details: set the variables in \code{revdeps/email.yaml}. You'll be prompted to
review the template before any emails are sent; or you can use
\code{revdep_email_draft()} to see a draft version.
}
\details{
To use this function, you'll need to give the gmailr app authority to
send emails from gmail. To revoke that authority, delete the \code{.httr-oauth}
file created in your working directory.
}
