% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{revdep_emails}
\alias{revdep_emails}
\title{Email reverse dependency maintainers about potential problems}
\usage{
revdep_emails(pkg = ".", revdeps = NULL)
}
\description{
Email reverse dependency maintainers about potential problems
}
\seealso{
Other revdep check functions: \code{\link{revdep_details}},
  \code{\link{revdep_report_summary}},
  \code{\link{revdep_results}}
}
