% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{revdep_env_vars}
\alias{revdep_env_vars}
\title{Environment variables to set for install and check processes while
running the reverse dependency check}
\usage{
revdep_env_vars(force_suggests = FALSE)
}
\arguments{
\item{force_suggests}{Whether to force the installation of the
suggested packages.}
}
\value{
Named character vector.
}
\description{
Environment variables to set for install and check processes while
running the reverse dependency check
}
