% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{revdep_maintainers}
\alias{revdep_maintainers}
\title{List maintainers of all reverse dependencies}
\usage{
revdep_maintainers(pkg = ".")
}
\arguments{
\item{pkg}{Path to package.}
}
\description{
List maintainers of all reverse dependencies
}
