% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{revdep_pkgs}
\alias{revdep_pkgs}
\title{Compute package revdeps}
\usage{
revdep_pkgs(pkg, dependencies = c("Depends", "Imports", "Suggests",
  "LinkingTo"), bioc = TRUE)
}
\arguments{
\item{pkg}{A character vector of package names whose reverse
dependencies should be checked. If \code{NULL} or empty, all CRAN
packages (and all Bioconductor packages if \code{bioc} is \code{TRUE}) are
returned.}

\item{dependencies}{Which types of revdeps should be checked. For
CRAN release, we recommend using the default.}

\item{bioc}{Also check revdeps that live in BioConductor?}
}
\value{
A tibble with a \code{package} column. If \code{pkg} contains
multiple packages, these are recorded in an additional column
\code{set} that allows you to match revdeps to elements of \code{pkg}.
}
\description{
\code{pkg_revdeps()} returns a tibble of reverse dependencies that you
can pass to \code{revdep_check()}.
}
\examples{
if (FALSE) {

# Compute revdeps of rlang
revdep_pkgs("rlang")

# Don't include Bioconductor revdeps
revdep_pkgs("rlang", bioc = FALSE)

# Compute revdeps of rlang, dplyr and purrr
revdep_pkgs(c("rlang", "dplyr", "purrr"))

# Return all packages of CRAN and Bioconductor:
revdep_pkgs(NULL)

# Return all packages of CRAN:
revdep_pkgs(NULL, bioc = FALSE)

}
}
