% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{revdep_report}
\alias{revdep_report}
\title{Markdown report of reverse dependency check results}
\usage{
revdep_report(pkg = ".", revdeps = NULL)
}
\description{
Markdown report of reverse dependency check results
}
\seealso{
Other revdep check functions: \code{\link{revdep_details}},
  \code{\link{revdep_emails}}, \code{\link{revdep_results}}
}
