% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{revdep_results}
\alias{revdep_results}
\title{Summary of reverse dependency check results}
\usage{
revdep_results(pkg = ".", revdeps = NULL)
}
\description{
Summary of reverse dependency check results
}
\seealso{
Other revdep check functions: \code{\link{revdep_details}},
  \code{\link{revdep_emails}},
  \code{\link{revdep_report_summary}}
}
