% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{revdep_todo}
\alias{revdep_todo}
\title{List packages remaining in the todo list}
\usage{
revdep_todo(pkg = ".")
}
\arguments{
\item{pkg}{Path to package}
}
\description{
List packages remaining in the todo list
}
