% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-loop.R
\name{run_event_loop}
\alias{run_event_loop}
\title{This is the event loop of the revdep check process}
\usage{
run_event_loop(state)
}
\arguments{
\item{state}{The full state of the check process:
\itemize{
\item \code{options} contains all check parameters.
\item \code{packages} is a data frame with the packages to check.
See details below.
}}
}
\description{
This is the event loop of the revdep check process
}
\details{
\code{state$packages} is a data frame with columns:
\itemize{
\item \code{package}: the name of the package
\item \code{state}: where we are with its check. Possible values:
\itemize{
\item \code{todo}: haven't done anything yet
\item \code{deps_installing}: the dependencies are being installed now
\item \code{deps_installed}: the dependencies were already installed
\item \code{downloading}: the source package to check is being downloaded
\item \code{downloaded}: the source package was downloaded
\item \code{checking}: checking with the old version right now
\item \code{checking-checking}: checking with both versions right now
\item \code{done-checking}: done with the old version, checking with the new
version right now
\item \code{checking-done}: checking with the old version, new version was
already done.
\item \code{done-downloaded}: done with the old version, check with new
version has not started yet
\item \code{done}: packages was checked with both versions
}
}

We only start the check with the new version after the check with the
old version, which simplifies the state transitions a bit.
}
\keyword{internal}
