% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-loop.R
\name{schedule_next_task}
\alias{schedule_next_task}
\title{Decide what to do next, from the current state}
\usage{
schedule_next_task(state)
}
\arguments{
\item{state}{See \code{\link[=run_event_loop]{run_event_loop()}} for a description.}
}
\description{
In we have reached the allowed number of workers, then we schedule an
idle job, we just need to wait until a worker is done.
}
\details{
Otherwise we schedule a job. In general the strategy is to finish check
as soon as possible, so if a package is in \code{deps_installed}, then we
schedule a check. Otherwise, if a package is in \code{todo}, then we
schedule a dependency install.

If there is nothing we can do now, then we schedule an idle job, i.e.
just wait until a worke is done.
}
\keyword{internal}
