% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_broken}
\alias{cloud_broken}
\alias{cloud_failed}
\alias{cloud_problems}
\title{Retrieve the names broken or failed packages}
\usage{
cloud_broken(
  job_name = cloud_job(pkg = pkg),
  pkg = ".",
  install_failures = FALSE,
  timeout_failures = FALSE
)

cloud_failed(job_name = cloud_job(pkg = pkg), pkg = ".")

cloud_problems(job_name = cloud_job(pkg = pkg), pkg = ".")
}
\arguments{
\item{job_name}{The job name, as returned by \code{\link[=cloud_check]{cloud_check()}}.}

\item{pkg}{Path to package.}

\item{install_failures}{Whether to include packages that failed to install.}

\item{timeout_failures}{Whether to include packages that timed out.}
}
\value{
A character vector with the names of broken packages, to be passed to \code{cloud_check()}.
}
\description{
Broken packages are those whose checks got worse with the dev version.
Failed packages are those whose cloud jobs failed, either because the spot
instance was shut down by AWS or because the checks used too much memory and
were killed.
}
\seealso{
Other cloud: 
\code{\link{cloud_cancel}()},
\code{\link{cloud_check}()},
\code{\link{cloud_details}()},
\code{\link{cloud_fetch_results}()},
\code{\link{cloud_plot}()},
\code{\link{cloud_report}()},
\code{\link{cloud_results}()},
\code{\link{cloud_status}()},
\code{\link{cloud_summary}()}
}
\concept{cloud}
