% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_browse}
\alias{cloud_browse}
\title{Browse to the AWS url for the job}
\usage{
cloud_browse(job_name = cloud_job(), package = NULL)
}
\arguments{
\item{job_name}{The job name, as returned by \code{\link[=cloud_check]{cloud_check()}}.}

\item{package}{If \code{NULL} browses to the URL of the overall job. If a package
name, browses to the URL for that specific package job.}
}
\description{
This is useful for closer inspection of individual jobs while they are
running or after the fact.
}
