% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_job}
\alias{cloud_job}
\title{Return the current cloud job}
\usage{
cloud_job()
}
\description{
This is automatically set by \code{\link[=cloud_check]{cloud_check()}} and only lasts for the current R session.
}
