% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_job_mapping}
\alias{cloud_job_mapping}
\title{Get a tibble of batch sub-job ids for all checked packages}
\usage{
cloud_job_mapping(job_name = cloud_job())
}
\arguments{
\item{job_name}{The job name, as returned by \code{\link[=cloud_check]{cloud_check()}}.}
}
\description{
Get a tibble of batch sub-job ids for all checked packages
}
