% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_plot}
\alias{cloud_plot}
\title{Plot the running time of a cloud job}
\usage{
cloud_plot(job_id = cloud_job())
}
\arguments{
\item{job_id}{The batch job_id, as returned by \code{\link[=cloud_check]{cloud_check()}}.}
}
\description{
Plot the running time of a cloud job
}
\seealso{
Other cloud: 
\code{\link{cloud_cancel}()},
\code{\link{cloud_check}()},
\code{\link{cloud_details}()},
\code{\link{cloud_fetch_results}()},
\code{\link{cloud_report}()},
\code{\link{cloud_results}()},
\code{\link{cloud_status}()},
\code{\link{cloud_summary}()}
}
\concept{cloud}
