% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_status}
\alias{cloud_status}
\title{Monitor the status of a cloud job}
\usage{
cloud_status(job_id = cloud_job())
}
\arguments{
\item{job_id}{The batch job_id, as returned by \code{\link[=cloud_check]{cloud_check()}}.}
}
\description{
Monitor the status of a cloud job
}
\seealso{
Other cloud: 
\code{\link{cloud_cancel}()},
\code{\link{cloud_check}()},
\code{\link{cloud_details}()},
\code{\link{cloud_fetch_results}()},
\code{\link{cloud_report}()},
\code{\link{cloud_results}()}
}
\concept{cloud}
