% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud.R
\name{cloud_summary}
\alias{cloud_summary}
\title{Display revdep results}
\usage{
cloud_summary(job_id = cloud_job(), pkg = ".")
}
\arguments{
\item{job_id}{The batch job_id, as returned by \code{\link[=cloud_check]{cloud_check()}}.}

\item{pkg}{Path to package.}
}
\description{
Displays nicely formatted results of processed packages run in the cloud.
}
\seealso{
Other cloud: 
\code{\link{cloud_cancel}()},
\code{\link{cloud_check}()},
\code{\link{cloud_details}()},
\code{\link{cloud_fetch_results}()},
\code{\link{cloud_report}()},
\code{\link{cloud_results}()},
\code{\link{cloud_status}()}
}
\concept{cloud}
