% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/async.R
\name{cran_compare}
\alias{cran_compare}
\alias{revdep_check_against_cran}
\alias{async_check_package}
\title{Compare local checks to CRAN checks}
\usage{
revdep_check_against_cran(
  dir,
  pkgs,
  num_workers = 2,
  flavour_pattern = "devel"
)

async_check_package(dir, pkg)
}
\arguments{
\item{dir}{The directory to perform local checks in.}

\item{pkgs}{A character vector of package names to check. If a
number, draw a random sample of CRAN packages of that size.}

\item{flavour_pattern}{A regexp to match against
\code{\link[rcmdcheck:cran_check_flavours]{rcmdcheck::cran_check_flavours()}}. If multiple matches are
found, the first element is used. This determines which CRAN
check to compare against.}

\item{pkg_name}{A package name.}
}
\description{
Compare local checks against CRAN checks. This can be useful for
running checks on a local patched version of R with the CRAN
results for that package. For this reason, the default value of
\code{flavour_pattern} is \code{"devel"}

The checks are performed in \code{dir} and results are saved to file in
that directory. You can call \code{revdep_check_against_cran()} with new
packages to test, old results are retained.

To reset the state, delete:
\itemize{
\item \code{checks/state.rds} which contains the results of completed checks.
\item \verb{cache/} which contains the crancache for dependencies.
}

The \verb{checks/} folders retains the \verb{R CMD check} files.
}
