% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{cran_revdeps}
\alias{cran_revdeps}
\title{Retrieve the reverse dependencies for a package}
\usage{
cran_revdeps(package, dependencies = TRUE, bioc = FALSE)
}
\arguments{
\item{package}{The package (or packages) to search for reverse dependencies.}

\item{dependencies}{Which types of revdeps should be checked. For CRAN
release, we recommend using the default.}

\item{bioc}{Also check revdeps that live in Bioconductor?}
}
\description{
Uses a hard coded CRAN mirror of \url{https://cloud.r-project.org} to ensure
that all users get the same results.
}
