% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{cran_revdeps}
\alias{cran_revdeps}
\title{Retrieve the reverse dependencies for a package}
\usage{
cran_revdeps(package, dependencies = TRUE, bioc = FALSE, cran = TRUE)
}
\arguments{
\item{package}{The package to search for reverse dependencies}

\item{dependencies}{Which types of revdeps should be checked. For CRAN
release, we recommend using the default.}

\item{bioc}{Also check revdeps that live in Bioconductor?}

\item{cran}{Should cran mirror be attached to getOpion("repos") if it
is not already present.}
}
\description{
Retrieve the reverse dependencies for a package
}
