% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db_setup}
\alias{db_setup}
\title{Make sure that the database exists}
\usage{
db_setup(package)
}
\arguments{
\item{package}{The name of the package under revdep cheking.}
}
\value{
Nothing
}
\description{
Make sure that the database exists
}
\keyword{internal}

