% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{pkgs_revdeps}
\alias{pkgs_revdeps}
\title{Compute package revdeps}
\usage{
pkgs_revdeps(pkg, dependencies = c("Depends", "Imports", "Suggests",
  "LinkingTo"), bioc = TRUE)
}
\arguments{
\item{pkg}{Path to package.}

\item{dependencies}{Which types of revdeps should be checked. For
CRAN release, we recommend using the default.}

\item{bioc}{Also check revdeps that live in BioConductor?}
}
\description{
\code{pkg_revdeps()} returns a tibble of reverse dependencies that you
can pass to \code{revdep_check()}.
}
\examples{
if (FALSE) {

# Compute revdeps of rlang
pkgs_revdeps("rlang")

# Don't include Bioconductor revdeps
pkgs_revdeps("rlang", bioc = FALSE)

# Compute revdeps of rlang, dplyr and purrr
pkgs_revdeps(c("rlang", "dplyr", "purrr"))

}
}
