% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{revdep_add}
\alias{revdep_add}
\alias{revdep_add_broken}
\title{Manually add packages to check}
\usage{
revdep_add(pkg = ".", packages)

revdep_add_broken(pkg = ".")
}
\arguments{
\item{pkg}{Path to package.}

\item{packages}{Character vector of package names to add}
}
\description{
Use \code{revdep_add()} to add a single package to the to do list. Use
\code{revdep_add_broken()} to add all broken packages from the last check
(this is useful if you think you've fixed the underlying problem in
your package)
}
