% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdepcheck.R
\name{revdep_check}
\alias{revdep_check}
\alias{revdep_reset}
\title{Run revdep checks}
\usage{
revdep_check(pkg = ".", dependencies = c("Depends", "Imports", "Suggests",
  "LinkingTo"), quiet = TRUE, timeout = as.difftime(10, units = "mins"),
  num_workers = 1, bioc = TRUE, warm_cache = FALSE)

revdep_reset(pkg = ".")
}
\arguments{
\item{pkg}{Path to package}

\item{dependencies}{Which types of revdeps to check}

\item{quiet}{Suppress output from internal processes?}

\item{timeout}{Maximum time to wait (in seconds) for \code{R CMD check} to
complete.}

\item{num_workers}{Number of parallel workers to use}

\item{bioc}{Also check revdeps in BioConductor?}

\item{warm_cache}{Warm the CRAN cache before performing the check?
Useful for fetching all dependencies (and, on Linux, also compiling them)
in a first run.}
}
\description{
\code{revdep_check()} is designed to work even if interrupted. If for some reason
you need to stop the checks, or the session dies, just rerun \code{revdep_check()}
and it will resume from where it last stopped. To completed reset the
results of a previous run, use \code{revdep_reset()}.
}
