% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{revdep_details}
\alias{revdep_details}
\title{Details of reverse dependency check results}
\usage{
revdep_details(pkg = ".", revdep)
}
\arguments{
\item{pkg}{Path to the revdep-checked package tree.}

\item{revdep}{Name of the reverse dependency.}
}
\description{
Details of reverse dependency check results
}
\seealso{
Other revdep check functions: \code{\link{revdep_emails}},
  \code{\link{revdep_report_summary}},
  \code{\link{revdep_results}}
}
