% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{revdep_email}
\alias{revdep_email}
\alias{revdep_email_draft}
\title{Notify revdep maintainers about problems}
\usage{
revdep_email(type = c("broken", "failed"), pkg = ".", packages = NULL)

revdep_email_draft(type = "broken", pkg = ".", data = email_data(pkg))
}
\arguments{
\item{type}{Type of problems to notify about; either "broken" (i.e. there
is a new \code{R CMD check} failure that did not currently occur) or
"failed" (i.e. the check failure either during installation or because
of a timeout).}

\item{pkg}{Path to package.}

\item{packages}{A character vector of package names. Use this if some emails
failed to send in the previous round. If omitted uses all packages.}

\item{data}{Optionally, supply a named list to provide your own parameters
to fill in the template}
}
\description{
This function uses gmail to automatically notify all maintainers of revdeps
that have failures with the new version of the package. The form of the
email is fixed, but it uses template parameters so that you can control
the details: set the variables in \code{revdeps/email.yaml}. You'll be prompted to
review the template before any emails are sent; or you can use
\code{revdep_email_draft()} to see a draft version.
}
\details{
To use this function, you'll need to give the gmailr app authority to
send emails from gmail. To revoke that authority, delete the \code{.httr-oauth}
file created in your working directory.
}
