% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{revdep_report_summary}
\alias{revdep_report_summary}
\alias{revdep_report_problems}
\alias{revdep_report_cran}
\alias{revdep_report}
\title{Markdown report of reverse dependency check results}
\usage{
revdep_report_summary(pkg = ".", file = "")

revdep_report_problems(pkg = ".", file = "")

revdep_report_cran(pkg = ".")

revdep_report(pkg = ".")
}
\arguments{
\item{pkg}{Path to package.}

\item{file}{File to write output to. Default will write to console.}
}
\description{
You can use these functions to get intermediate reports of a \code{\link[=revdep_check]{revdep_check()}}
running in another session.
}
\details{
\code{revdep_report()} writes the \code{README.md} and \code{problems.md}. This is
normally done automatically when the checks are complete, but you
can also do it when checks are in progress to get a partial report.
}
