% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{revdep_report_summary}
\alias{revdep_report_summary}
\alias{revdep_report_problems}
\alias{revdep_report_failures}
\alias{revdep_report_cran}
\alias{revdep_report}
\title{Markdown report of reverse dependency check results}
\usage{
revdep_report_summary(pkg = ".", file = "", all = FALSE, results = NULL)

revdep_report_problems(pkg = ".", file = "", all = FALSE, results = NULL)

revdep_report_failures(pkg = ".", file = "", results = NULL)

revdep_report_cran(pkg = ".", file = "", results = NULL)

revdep_report(pkg = ".", all = FALSE, results = NULL)
}
\arguments{
\item{pkg}{Path to package.}

\item{file}{File to write output to. Default will write to console.}

\item{all}{Whether to report all problems, including the ones that
were already present in the old version of the package. This potentially
generated a lot of output, most of which was irrelevant, so they are
omitted by default, and only problems seen with the new version of
the package are reported.}

\item{results}{Cached results from \code{db_results()}. Expert use only.}
}
\description{
You can use these functions to get intermediate reports of a \code{\link[=revdep_check]{revdep_check()}}
running in another session.
}
\details{
\code{revdep_report_summary()} writes the contents of \code{README.md}, by
default to the console. This is handy to quickly inspect the (current)
list of problematic packages.
}
