% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-loop.R
\name{run_event_loop}
\alias{run_event_loop}
\title{This is the event loop of the revdep check process}
\usage{
run_event_loop(state)
}
\arguments{
\item{state}{The full state of the check process:
\itemize{
\item \code{options} contains all check parameters.
\item \code{packages} is a data frame with the packages to check.
See details below.
}}
}
\description{
This is the event loop of the revdep check process
}
\details{
\code{state$packages} is a data frame with columns:
\itemize{
\item \code{package}: the name of the package
\item \code{state}: where we are with its check. Possible values:
\itemize{
\item \code{todo}: haven't done anything yet
\item \code{deps_installing}: the dependencies are being installed now
\item \code{deps_installed}: the dependencies were already installed
\item \code{checking}: we are checking the package right now
\item \code{done}: packages was checked, result is in the database
}
}
}
\keyword{internal}
